Move business related logic to business layer
